fix: apply current project to list datum spec
This fixes the broken behaviour of list datum from file, where a user
does not provid the project in the PPS Input section, and we wrongly
infer the default project. Now, we use the current project in the local
context to infer the project.

Solves CORE-1507, which unblocks the 2.5 release GA release.