[PFS-37] Add an RPC Endpoint to PFS to Search for a File in Branch  (#8606)
This endpoint takes a starting commit and a filepath and serially searches through the branch commit-by-commit. If a reference to the filepath, we add the commit to a list to return to the user. If the search request hits its deadline, it returns what it found so far as well as the last commit it searched. This allows a user to start a subsequent searching from that commit.

This endpoint is primarily designed to be used by Console to display a list of commits to a user that contain a modification to a specific file. The goal is to have elements in the user interface that allow a user to continue a search (or make it easy for Console to continue a search in the background as it updates the list.)