Compact Diff File Sets in Old Commits When Calling Find Commits (#8704)
This PR implements a simple backfill strategy for handling reads on uncompacted diff file sets. These file sets will be compacted at read time by the PFS master's driver. It is possible that multiple concurrent client calls could lead to the driver attempting to recompact the diff file set for the same commit at the same time and waste work, but we think its a worthy trade off for keeping the master code simpler.

To make this more efficient, the PFS team revisited the IsCompacted() logic which involved making optimizations to Flatten() so that it would stream over a set of primitives rather than holding all the primitives in memory. Some helper functions were also modified to reduce duplicate code.