Adjust mount-server auth flow to authenticate pipeline extension (#8765)
Updates the auth flow in the Jupyter extension to accommodate authentication when running
in sidecar mode. Previously, the pipeline extension only works unauthenticated or with the
mount-server running in subprocess mode because the Jupyter extension itself did not have
access to the session token that the mount-server negotiates upon login when the processes
do not share the same container. This change updates the auth flow so that the mount-server
auth is done in two steps, returning a redirect to the login page in the first and returning the
Pachyderm configuration along with the session token in the second step. This enables the
Jupyter extension to write the same configuration into its own container and use those for
creating pipelines in Pachyderm deployments with authentication enabled.

---------

Co-authored-by: Kevin Yang <kevin.yang@hpe.com>