Populate empty projects in project selection within Jupyter extension (#8855)
Currently, the way that we populate the projects list in the Jupyter extension is by taking the repos list and getting all projects from each of the repos. The issue with this is that projects with no repos can exist, and it can cause confusion if a user creates a new project without repos since it won't show up in the extension. This change adds a /projects endpoint to the mount server and makes the extension query it to retrieve all projects, including those without any repos.

---------

Co-authored-by: Kevin Yang <kevin.yang@hpe.com>