Enhance jupyterlab_pachyderm plugin to support non-prived mode
HPC container support to enable a secondary path within the
jupyterlab_pachyderm plugin to enable use in a container without --privileged.
This support is currently enabled by setting the environment variable NONPRIV_CONTAINER.
In the non-privileged mode, the mount-server is launched in an unshared namespace
and mounts to an alternate directory (/mnt/pfs), and the /pfs directory is replaced
with a softlink to the the /mnt/pfs (via the /proc filesystem).