storage/kv: port obj.Client utilities (#8643)
We would like to eventually migrate away from our `internal/obj` package in favor of the [go-cloud sdk](https://pkg.go.dev/gocloud.dev) interfaces.  We already use the Go CDK clients for imports and export done with PutFileURL and GetFileURL, but we would also like to use them for the main storage backend.  The original motivation for this was better object metadata when listing the contents of a bucket; it was easier to switch to the Go CDK interfaces for import/export than to add that functionality to our own interfaces for all the implementations.

This PR ports a lot of the extra functionality that we have for the `obj.Client` interface to the `kv.Store` interface.   The `kv.Store` interface is used throughout the storage layer since we only deal in chunks.