Correct message typing
The actual Cluster Defaults message is a properly-typed protobuf; it
defines the schema the user-submitted cluster defaults must follow.

The user-provided input is JSON, and hence a string.  It is stored
verbatim, and unmarshaled when needed.

The client SHOULD validate that user input unmarshals; the server
MUST.  But those are in another story.