Fix move branch trigger test (#9146)
The move branch trigger test was relying on the branch movement
operation completing before the asynchronous trigger evaluation to be
successful. Recent improvements to trigger performance made this race
condition more apparent. Also, the test didn't really reflect the
expected behavior after the 2.6 provenance changes. The test was updated
to account for both of these issues.