[2.7.x] Consistently format full timestamps in pachctl (#9176) (#9191)
This gets all the places where we print times from protos and formats
them as RFC3339 without nanoseconds.

In the past this printed nanoseconds by coincidence, but I don't think
anyone needs them and it takes up 9 chracters of space to do so. (Those
that want them can run --raw to get it in that form).