Djanicek/core 1830/det exemple test (#9171)
This PR adds a test that starts up pachyderm with determined and creates
a `determined` pipeline. It then verifies the new service user is
created in determined by the pachyderm command.

 **---Helm chart change---** 
One thing I noticed when running in parallel, is that doing a helm
install in different namespaces failed. This appears to be due to the
Priority Classes in the determined chart. Priority classes are
non-namespace so `helm install` fails when it tries to make the same
class with different namespace metadata, I added some logic to not
create those Priority Classes if they already exist.