[CORE-1778] pachctl create pipeline --dry-run (#9205)
This PR adds a `--dry-run` flag to `pachctl [create | update] pipeline`.
When set, the user spec is merged into the cluster defaults to form an
effective spec, which is returned and printed, but the pipeline is not
actually created/updated. This enables a user to see the results of a
creation or update without actually changing the state of the system.

It also adds formatted and raw output on dry run.