[CORE-1708] Propagate cluster defaults (#9020)
This PR implements propagation of cluster defaults to pipelines when the
`regenerate` flag is true. When it is, the effective spec for each
pipeline is calculated from the cluster defaults and the user spec; if
the newly-calculated effective spec differs from the old effective spec,
then the effective spec is replaced (as in an update pipeline
operation). Furthermore, if the `reprocess` flag is set then the
pipeline will reprocess datums (again, as in an update pipeline
operation).

In order to implement this, create-pipeline transactions had to be
reworked. They were previously just the create pipeline request message;
they are now a special message which captures the user and effective
specs in addition to the create pipeline request. Rather than implement
this change in the old private `ppsBuilderClient.CreatePipeline` methods
— uncalled in the codebase — the dead code is deleted.

The defaults-handling code is abstracted, in order to provide a single
place for wrapping and unwrapping of JSON values and in hopes of
improving legibility of the API server.

The former `details_json` name is updated to the more accurate
`user_spec_json`, to pair with the new `effective_spec_json`.

---------

Co-authored-by: Daniel Janicek <116583733+djanicekpach@users.noreply.github.com>
Co-authored-by: Albert Cui <albert@pachyderm.io>