pachctl: misc generate-download-url: show a full clickable link if the server knows its address (#9218)
A while back we added a proxy_host and proxy_tls field to
InspectCluster, populated from the helm chart. With that, we can
generate the full download URL from `pachctl misc
generate-download-url`:

```
$ go run ./src/server/cmd/pachctl misc generate-download-url images@master:/
http://192.168.1.70/archive/ASi1L_0EAIEAAGltYWdlc0BtYXN0ZXI6LwDjg9sf.zip
```
If the server isn't around and doesn't have that field set, we just
output the `ASi1L_...` part as we did before.