Djanicek/ci connection retries (#9227)
Flakes have been building up in the deploy tests for a while to the
point where they basically never pass. There were actually 3 or four
different issues in these tests. Since one of those is dex/auth errors
related to CORE-1726, I'm fixing that issue and doing some targeted
cleanup to try to stabilize the tests. Some of these changes need
follow-ups and have jira tickets linked.

There are still some issues with flaky mockIdp, as well, but for the
sake of keeping this moving I've added those to a ticket.
[CORE-1922](https://pachyderm.atlassian.net/browse/CORE-1922)

**Changes by file:**

_.circleci/main.yml_

This more accurately distributes CI executor resources as needed. Since
arm64 runs fewer, smaller tests, I'm beefing up amd64 so it has more
memory/cpu to work with.

_etc/helm/pachyderm/templates/proxy/pachd-proxy-backend-service.yaml_

This hopefully helps with disconnects on the backend between envoy and
pachd since we often see disconnect errors in tests that go away on
retry. [Slack
link](https://hpe-aiatscale.slack.com/archives/C057E0BPKQB/p1692638535321299)

_src/internal/testutil/auth.go, src/internal/testutil/identity.go_

This is basically CORE-1726. The mockIdp is flaky, so including retires
helps prevent the dex/token expired errors.

_src/testing/deploy/upgrade_test.go_

This is the real flaky test culprit. The current theory is that now due
to performance improvements, the putfiles (when not waiting for commits
to finish) are triggering too many times. in my testing, I confirmed
that adding the wait consistently triggers commits 10 times. Not waiting
gets an extra 5 commits.

Additionally, having 10+10 calls to `putFile` generates 21 commits in
the input repo because of the empty commit, so we never trigger the last
one. Bumping to 11 commits in the current version consistently passes
here, but the correctness is not clear.
[PFS-135](https://pachyderm.atlassian.net/browse/PFS-135) exists to
validate the correctness of this change as it's currently unclear.

[PFS-135]:
https://pachyderm.atlassian.net/browse/PFS-135?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[CORE-1922]:
https://pachyderm.atlassian.net/browse/CORE-1922?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ