Djanicek/core 1862/perf improvements (#9192)
When I initially made the test export for Grafana I moved all of the
heavy lifting into the pachyderm side so it did not hold up the CI
pipelines. This was good, but I also assumed that pachyderm.pachops.com
would have no trouble keeping up with the demand. Even if it fell behind
on processing the data and exporting it to Postgres, it would be able to
catch up over night or on the weekends when demand was low.

A few weeks ago we had a busy week and demand did not catch up over the
weekend. Since then it has been slowly getting more and more behind.
This PR hopefully puts it in a better position. There are a few fixes:

1. Insert the test result into Postgres in batches now that we are on
sqlx 1.3 so we don't send individual inserts to postgres.
2. Put the postgres upload into goroutines so large files can upload as
we read the file.(Note, our cloud-sql instance has a capacity of 100
connections and the rest of the system is taking about 20, so 10
_should_ be ok)
3. Increase the parallelism spec for the go-test-results-sql-egress
pipeline so we can run on more than one datum at once.
4. In the collector that sends data from ci, upload both required files
as one datum with ModifyFileClient. This cuts down on the number of
datums pachyderm has check and skip over.

With all this I think we should be in better shape. If after this we
still are falling behind I think we would want to increase the pipeline
parallelism spec, and increase the available cpu and memory in this
cluster.

Local testing shows the upload of a large(1.25GB) file of test results
went from 30-45 seconds to upload down to 8-12 seconds to upload.
Additionally, while uploading 4 files it was taking 5-6 minutes locally,
running the same test is now down to roughly 1-2 minutes. For small
files locally the upload of 4 files is now finishing in 15-30 seconds,
so hopefully we see similar gains in the real cluster.

This upgrades to sqlx 1.3.5 in the even that we are not already
upgraded.