Djanicek/auth debug (#9247)
This updates k8s to the latest minikube-supported version: 1.27.4.
https://github.com/kubernetes/minikube/releases

It turns out some in CI requests were sometimes being routed to the
wrong pods/namespaces. The current theory is this was a bug in k8s since
we were using an old version, and the upgrade helps. In testing, in 10
runs of the deploy tests we've had 1 failure, in it was unrelated to the
dex or disconnect errors that have been plaguing us.

This requires some updates since PodSecurityPolicy is now deprecated and
a couple Kubernetes resources moved from v1beta to v1.