Update jupyter_server version to 2.7.0 to match Determined AI
Determineds task environments have jupyter_server==2.7.0, but
jupyterlab-pachyderm is locked to  jupyter_server>=1.6,<2 which
prevents it from being installed into a Determined task environment.
Change the locked version to jupyter_server>=2.7.0,<2.8  to match Determined
as of 0.24.0, yet allow for fixes.  jupyter_server recommends that the version be pinned
to avoid plug-in breakage, so this locks major and minor.

The integration tests use an unauthenticated interface between the
tests and the dev_server.   Add an jupyter_server identity provider
that disables authentication to enable this to continue working with
jupyter_server >= 2.0.

Also add pytest-jupyter to the options.extras_require which is used
by the extensions test.  It is a new dependency of the latest
jupyter_server version.

Create the mount_server.log in CI such that the tail of it works.
Previously, the tail -f fails because the file did not exist until
after pytest starts.