[Python-SDK] [2.7.x] Catch serialization errors (#9216) (#9255)
We now catch serialization errors and convert the unintuitive error
message into something more understandable to the average user. Example
(description field has a value to `True`):
```
TypeError: An error occurred while trying to serialize the following CreateRepoRequest message.
  This is most likely due to one of the fields of this message having a value with an incorrect type.
	Message: CreateRepoRequest(repo=Repo(name='sertest', type='user', project=Project(name='default')), description=True, update=False)
```

In writing the test for this feature I discovered that the current
interceptor didn't catch errors from a streamed response. In fixing
this, I also discovered that catching is connection error from a
streamed response is flaky but I could not find a more robust method for
doing so.