proxy: discover endpoints before k8s health check passes (#9235)
This change publishes k8s endpoints for pachd/console pods as soon as
they start, instead of as soon as they pass the readiness check. This
avoids 3 loops; k8s polling for readiness, envoy polling for DNS records
existing, envoy polling the readiness. Instead, the envoy DNS poll
returns endpoints as soon as the pod starts, and then it starts health
checking the pod every 1s. This should make traffic route to the proxy
much sooner.

I also turned down DNS error handling polls to a more reasonable
interval.

Finally, I adjusted the panic threshold so that unhealthy backends are
unlikely to be selected. (Envoy enters "panic mode" when a certain % of
backends are unhealthy; in panic mode, health is ignored. The idea is to
prevent cascading failure when under too-high load (broken backends get
some of the load this way), but I don't think we care about that.
Rather, we'd prefer additional load on the healthy replica over
returning an error response.)