Warn when stopping/deleting a pipeline that doesn't exist (#9284)
Internally, we like to swallow errors matching `errutil.IsNotFoundError`
when deleting or stopping pipelines. For pachctl, it seems that users
want to know when they have mistyped something. This adds an option to
the API to not swallow those errors, and turns that option on for
pachctl.

The error messages are somewhat hard to read ("branch not found"), but
#9273 fixes that. The changes are safe to merge in either order; this PR
is not based on that branch.