pfs/pps: make everything take a context (#9283)
This is the promised refactoring that plumbs in a context to all public
PFS/PPS APIs (where they call each other).

This is necessary for the database rework; the parts that don't take
contexts happen to be parts that aren't moved to the new database yet.
I'm adding it because it allows us to do some input validation that
wasn't possible for (in a ... past PR; I'll rebase that one on top of
this once approved).

The most non-mechanical part of this was making constructors of
propagators NOT take a context, but instead making the run method take
that. That eliminates a context inside a struct, which is fortunately
decreasing in number rapidly in the codebase!

The auth API still doesn't take contexts, even though it's basically the
same idea as PFS and PPS in terms of its interface. That's because auth
is last on the list of things to move over to the database, and there is
some talk about changing it dramatically in the near future. If that
happens, we might as well not bother with the refactor yet; nothing will
need it in this code's lifetime, it seems.