circle: use all available cores for test-go (#9294)
I think GOMAXPROCS=8 predated having pgbouncer in the tests. Let's see
if this speeds things up.

This also lets me run the tests locally without setting GOMAXPROCS
(which is 32 on my machine).

I tested bumping up the instance size to 2xl; that makes the tests about
a minute faster, so probably not worth it. On my local machine with 32
cores, it's not much faster, docker-proxy is using all the CPU, and
postgres and pgbouncer are also quite busy. So I think this is fine.