Generate documentation from protos (#9271)
This adds protoc-gen-doc to `make proto`, outputting JSON and Markdown
documentation.

The JSON could be used as input when the friendly docs are being
created, so it can be maintained here but displayed on
docs.pachyderm.com. The Markdown is just nice if you're browsing our
repo.

One note: the fields are output in "proto" format and not "JSON" format.
We would like users writing a CreatePipelineRequest to use the JSON
format, so we might need to adjust that during doc generation.