Allow only protojson field names in example pipelines (#9304)
This deprecates the use of "snake_case" and string integers in the
examples. pachctl still allows them and that will not change.

I also ran 'deno fmt' on all the files that can be autoformatted (our
proprietary `{} --- {}` format doesn't autoformat, that's actually YAML
not JSON).

Finally, I added a CI rule to ensure that every `.pipeline.json` meets
the spec. That way if you change the protos, you will know what
pipelines you break!

This also satisfies CORE-1772 enough for me. That ticket wants us to run
`pachctl validate` but `yajsv` is much faster and is 90% as good, so I'm
going to close it.