pfs: adds WorkerEnv, WorkerConfig, and Worker (#9302)
Separates PFS worker code into its own type. The PFS worker will only
run in pachws, and full pachds in production, but is also setup by real_env during tests.

WorkerEnv only contains a TaskService, DB, and ObjClient. It doesn't need the
rest of the PFS DAG, the listeners, or the Auth Service. In order to
break the dependency on the Auth Service, the GetFileURL task had to be
changed to deal with Filesets, rather than Commits. As a general rule,
Filesets should be passed around to refer to data, not Commits.