Move PFS Load Tests to Debug Server (#9253)
These changes move the PFS Load Test methods to the debug server. The
motivation is to remove the pach client as a dependency of PFS.
Currently the way a service does a self RPC is by importing a pach
client, which also means it has to depend on all of the APIs.

The Debug server is a good place to move this functionality because it
is, and always will be, at the top of the dependency graph (without
anything depending on it).
Adding the Task Service as a dependency of the Debug Server makes sense
because it would allow us to parallelize our collection of debug dumps.