pachctl: Add a command to decode protos (#9285)
Sometimes you see marshaled protos floating around, like in the
database. Now pachctl can decode them.

```
$ pachctl misc decode-proto pfs_v2.BranchInfo 0x0A1B0A120A03666F6F1204757365721A050A036261721205656467657312530A1B0A120A03666F6F1204757365721A050A0362617212056564676573122064326362653833333161616634623634396532363430643566306662336165311A120A03666F6F1204757365721A050A03626172
{
  "branch": {
    "repo": {
      "name": "foo",
      "type": "user",
      "project": {
        "name": "bar"
      }
    },
    "name": "edges"
  },
  "head": {
    "repo": {
      "name": "foo",
      "type": "user",
      "project": {
        "name": "bar"
      }
    },
    "id": "d2cbe8331aaf4b649e2640d5f0fb3ae1",
    "branch": {
      "repo": {
        "name": "foo",
        "type": "user",
        "project": {
          "name": "bar"
        }
      },
      "name": "edges"
    }
  }
}
```
(No it's not messed up, I have a branch named "edges" in my instance and
happened to pick that.)

It understands hex (with or without a 0x prefix), base64, and raw
binary. You can put the data on the cmdline, or you can pipe it in from
stdout. Without args, it prints all message types it knows about.

This change broke the linter (it doesn't think `min` is a builtin), so I
upgraded it. The linter upgrade unfortunately upgrades depguard, whose
config format changed. I fixed it (and made sure it actually worked).