[CORE-1972] refresh autoscaling goroutines when pipelines update (#9330)
This PR updates a test to show a customer issue. It adds auth and
autoscaling to the test and waits for the pipeline to run before
updating.

The core issue is that the goroutines used to monitor autoscaling are
not restarted with the latest pipeline info / auth creds when the
pipeline is updated. This PR fixes that.

---------

Co-authored-by: acohen4 <alonco124@gmail.com>