use unique names and clear DB for test (#9354)
In a funny twist this test fails exactly 50% of the time. It always
fails on the first run, then passes when gotestsum runs it
`ValidDetails` on it's own.
There are 2 flakes:
* Using the same name for a pipeline gives an error that it exists
already
* Not removing the old pipeline causes the affected pipeline length
check to get 2 pipelines instead of 1

The delete all should solve both of these, but I it's just good practice
to use unique names per test, both to avoid false positives/negatives
and to help anyone who might add tests in the future.