Update jupyter_server version to 2.7.0 to match Determined AI (#9351)
# Summary

Determined's task environments have jupyter_server==2.7.0, but
jupyterlab-pachyderm is locked to jupyter_server>=1.6,<2 which prevents
it from being installed into a Determined task environment. Change the
locked version to jupyter_server==2.7.0 to match Determined as of
0.24.0. jupyter_server recommends that the version be pinned
to avoid plug-in breakage.

The integration tests use an unauthenticated interface between the tests
and the dev_server. Add an jupyter_server identity provider that
disables authentication to enable this to continue working with
jupyter_server >= 2.0.


# Testing

Verified the plug-in mount capability with this configuration.
Are there other specific areas that we need to verify?

`docker run -it -p 8888:8888 --cap-add SYS_ADMIN --device /dev/fuse
pachyderm/notebooks-user:dev`


![image](https://github.com/pachyderm/pachyderm/assets/84593277/aecfb5d8-31bc-485c-ae5f-2a5c5b0c6830)