handle race between delete repo and set commit (#9347)
This change should fix the TestSpoutPachctl fix. The full [design is
here](https://www.notion.so/2023-09-01-TestSpoutPachctl-Flake-mini-design-0bfa96b7a8e445889c18722d9c719c34).
This fix corresponds to option 3 in the design: verify the repo exists
before updating commit totals and diffs.

This should resolve the bug with normal pipeline deletion as well as
delete all, but it does not settle on a solution to Option 6.(That is
upcoming in a deep dive - to be scheduled)

I've been thinking about where to put this check, and I think this
location makes sense, but I'm definitely open to other ideas here. It
could go higher up so we only check once for both setDiff and setTotal,
but that does open up the possibility of the repo being removed between
the two sets.