enterprise command updates (#9334)
Minor edits and some examples added to the enterprise CMDs for docs. 

One thing I did notice that was interesting, however, is that when you
deactivate enterprise via `pachctl enterprise deactivate` you can still
use auth to log in. Is this known/expected and does it freeze the
ability to create new users?