internal/client: add functions for file methods (#9372)
There are many file utility methods on the pach client, and they are
pretty widely used. But in order to take advantage of them, you have to
pay the cost of the pach client abstraction, which means a dependency on
all the services.

These changes create functions for each of these methods and change the
methods to call into the functions, so they are useful anywhere a
pfs.APIClient is available.