Find the appropriate mount-server pid on HPC systems
Singularity/Apptainer/Enroot container runtime platforms provide
much less isolation than docker.   The pgrep command inside the
container may find the pid of mount-server instances running inside
other containers on the same node -- even by a different user.

This change limits the search for mount-server processes that are in
the same session as the command we launched.   We get the pid
of the command we launch, but it is wrappered by a bash and an
unshare command before the mount-server we are trying to find.