Add the ability to run Starlark during debug dumps, and run debug dumps locally (#9338)
This adds a (placeholder) ability to run Starlark during debug dumps,
and a pachctl command to run the dump locally. I also fixed a formatting
bug with "pachctl debug template", which was using the wrong YAML
encoder.

This is part 2 of the Starlark debug dumps project; part 1 was adding
the Starlark infrastructure, and part 3 is actually adding useful
functionality to Starlark debug dumps.

For developers developing scripts, use `starpach run -p debugdump` or
`starpach shell -p debugdump` to develop them. Then put the script in a
template, give it to a user, and have them run `pachctl debug local
template.yaml`. To have a script always run, add it to
`src/server/debug/server/debugstar/starlark` and it will be built into
pachd and pachctl for use with the `builtin` option in the template.
(They are run by default on the server.)