Dynamic k8s API for starlark scripts (#9350)
This adds a Kubernetes library for starlark scripts. It uses the dynamic
k8s client, which is the least horrible option. (The other options were
to use reflection against the static client, or generate a starlark
client from the OpenAPI specs.) I didn't use reflection because the
static client can't access dynamic resources; say we want to dump Istio
configs, you can't do that with the static client. I didn't do the
codegen because I thought it would take too long, but in the end not
doing it took too long. (We'd have to ship the compiler, though, for the
same Istio reason above.) Dynamic is the least bad, but it's not great.

It's actually amazing how bad the Kubernetes API is. I used to think
they had a good reason for not using protos + grpc, but nope, they did
not. Instead they just made a pretty bad implementation of grpc (oh but
with versions!) Looking forward to version 2.0.

The dynamic client is horrible in many regards. We have to figure out
that Kubernetes has a thing called "pods" by querying the server. That's
literally not compiled into the client anywhere!!! These methods do not
take contexts (and the bug for this is closed with "who cares why would
that need a context?"), so creating a k8s starlark client can take
unbounded time. Unfortunate, but actually very difficult to fix.

Kubernetes also has the concept of "subresources", which pod logs are
one of. The dynamic client has no clue how to handle these; it has code
to handle them, but it gets it totally wrong.

The library is designed to keep charging along in case of errors; error
objects have the same API surface as the result objects, and when the
debug dumper sees them it writes an error file instead of the requested
file. This is necessary because starlark errors are fatal; the script
cannot continue when there's an error.

To get a taste of the API, here's a script that dumps pipeline RCs as
JSON, and all associated pods:

```python
def dump_rcs():
    rcs = k8s.replicationcontrollers.list(labels = {"suite": "pachyderm", "app": "pipeline"})
    for rc in rcs:
        name = rc["metadata"]["name"]
        dump("%s/rc.json" % name, json.encode(rc))
        pods = k8s.pods.list(labels = rc["spec"]["selector"])
        for pod in pods:
            podname = pod["metadata"]["name"]
            dump("%s/pod-%s.json" % (name, podname), json.encode(pod))

if "replicationcontrollers" in dir(k8s) and "pods" in dir(k8s):
    dump_rcs()
```

You can also open up a shell and write something like `pods = k8s.pods`
and `print(pods)` or `dir(pods)` to see the functions available.

Finally, if you are wondering about the starlark.go change that sets
`ReassignGlobals` to true... it turns out no code reads the other fields
in that struct; if it's false, then you can't have top-level for loops
or if statements 😂. We are totally OK with those things, so we enable
them here.