Add a /pfs HTTP endpoint that allows browsing files stored in PFS (#9187)
This lets you treat Pachyderm like a webserver; generating HTML in
pipelines and then viewing them in the browser:


![Capture](https://github.com/pachyderm/pachyderm/assets/2367/0934f3f3-ebc0-4e48-ba2f-e47ffaaf6943)

There are also basic HTML directory listings:


![Capture](https://github.com/pachyderm/pachyderm/assets/2367/4d904e71-b692-4bd1-851c-0e328a2ddfa7)

The HTTP supported is quite extensive; you can download parts of a file
in parallel with `Range` headers, and use conditional requests like
`If-None-Match` or `If-Not-Modified`. `HEAD` requests are also
supported. Caching headers are provided so that your browser knows it's
safe to cache immutable files forever, but not cache mutable files (in
open commits) at all.

Some details:
* there is more to do here, like adding pages that list projects, repos,
branches/commits, etc. so that you don't have to guess a URL to get
started.
* we need to OAuth so people don't have to pass authn-token in the URL
* we need to implement PUT for uploading files