[FRON-1065]: Add RerunPipeline RPC (#9287)
This change is in support of the [rerun pipeline console
feature](https://pachyderm.atlassian.net/browse/FRON-1022). Some
research and implementation ideas can be found
[here](https://www.notion.so/Rerun-SubJobs-in-console-5bef4a4d328f444886fee5e423785355).

This PR adds an RPC to the `pps.proto` that when called takes a pipeline
and the option to reprocess. The backend then retrieves the last
pipeline spec and does a pipeline update resulting in a rerun of the
pipeline.
```
message RerunPipelineRequest {
  Pipeline pipeline = 1;
  // Reprocess forces the pipeline to reprocess all datums.
  bool reprocess = 15;
}
```

We want to add a dedicated rpc to pachyderm to handle this operation for
a few reasons.
- Only needing to make one rpc call to complete this operation
- Not needing to retrieve and pass around the input spec in console
- Simplify the arguments needed to run this operation

Our biggest concern is adversely modifying the user's input spec when
trying to rerun a pipeline using the existing `CreatePipelineV2` RPC.

---------

Co-authored-by: Jonathan Rockway <2367+jrockway@users.noreply.github.com>
Co-authored-by: Jeff Rifwald <jmcriffey@gmail.com>