[CORE-1956] Remove repeating header when listing files (#9405)
Currently, `pachctl list file` and `pachctl diff file` print a
variable-width format, and repeat the header lines periodically within
the body text. This is unexpected behavior; this PR changes it to a
constant-width format. It also eliminates the old dir/file type column;
the same information is indicated by the trailing slash in the name.

Old format:

```
NAME              TYPE SIZE
⋮
/wikipedias                           file 2B
/wikiquote                            file 3B
/wikisource                           file 3B
/wikispace/                           dir 3B
NAME              TYPE SIZE
/wikispecies      file 3B
```

New format:

```
SIZE    NAME
⋮
2B      /wikipedias
2B      /wikiquote
2B      /wikisource
2B      /wikispace/
2B      /wikispecies
````


The first column is now SIZE, and is padded to allow for numbers and
units. The second column is NAME, and is the file name. It ends in “/”
when the file is a directory.

Removed `withCommit` because it was never true; removed `fullTimestamps`
because it was only ever used when `withCommit == true`.