[CORE-1984] Switch from UnimplementedAPIServer to UnsafeAPIServer (#9409)
*This is no a customer-affecting change.*

The Go gRPC bindings represent an API service as an interface with one
method for each gRPC RPC. In order for a type to implement the
interface, it must implement each method in the interface. For each
service, the bindings generate _two_ possible embeddable types:
`UnimplementedAPIServer` or `UnsafeAPIServer`. These two are arguably
misnamed: `UnimplementedAPIServer` implements every Go-level method
required by the service interface (returning a runtime status code
indicating an error) in hopes that the programmer will override it; on
the other hand `UnsafeAPIServer` does not implement Go-level methods,
meaning that compilation fails unless they are implemented by the
programmer — i.e., a service embedding `UnsafeAPIServer` is compile-time
typesafe and doesn’t raise runtime errors (it does handle unimplemented
RPCs at the gRPC level, returning the same error that
`UnimplementedAPIServer` does. As an aside, perhaps it is appropriate
that this story is CORE-1984: `UnimplementedAPIServer` implements all
methods, and `UnsafeAPIServer` is safe, in the sense that the programmer
cannot accidentally leave a method unimplemented.

`UnimplementedAPIServer` is useful when implementing a service defined
in a different module, in order to prevent having to increment the other
module’s major version every time an RPC is added, but is unnecessary
when implementing a service defined in the same module.
`UnsafeAPIserver` is useful when implementing a service to ensure that
the programmer has not forgotten to implement all RPC methods. In our
case, we don’t envision implementations of our gRPC API services outside
of this repo, and we generally do want to ensure that we don’t
accidentally miss an RPC.

More details [here](https://github.com/grpc/grpc-go/issues/3669).

Of note, this PR does not touch PFS.