More trigger fixes (#9246)
This PR fixes a few issues related to triggers that were identified by
issues with TestUpgradeTrigger. The first fix is to make triggers
idempotent. Triggers are evaluated / fired during commit finishing and
commit finishing must be idempotent since it isn't guaranteed to only
occur once for a particular commit (PFS masters acquire a distributed
lock that doesn't provide strict mutual exclusion). Since triggers were
not idempotent, they could be fired more than once for a particular
commit. The second fix was to make the commit count trigger ignore the
initial empty commit that is created when a branch is created. This just
isn't behavior that a user would expect. The third fix was to update the
TestUpgradeTrigger test to do an additional commit to account for the
off-by-one error in the old versions, wait for the triggers to fire for
the existing commits before upgrading, and count the commits created by
the triggers to ensure that the correct number of commits are created
(before the idempotent change, there would be more commits than expected
here). There is still some investigation needed for why commit finishing
is firing multiple times because that should be a rare edge case when
the distributed lock is lost.