dockertestenv: expose configuration for direct and pgbouncer connections (#9459)
Trying to setup postgres listeners in NewTestPachd. I need access to a
direct and indirect connection for the same database. There was a bunch
of duplicated code in dockertestenv, so I restructured it around setting
up multi-db environments (direct and PGBouncer), with helper functions
for getting one DB vs. the other, or converting the configuration to the
various option formats used through the codebase.